+++ changelog.md
@@ -64,3 +64,4 @@
 * Error message for wrong range more verbose
 * Refactor logging
 * Remove unnecessary code
+* Tests small refactor